CHROMIUM: video: tegra: save/restore raster timing parameters
The detach sequence needs to send 2 dummy frames to activate register
values. Setting the timing parameters to minimum for speed up the
detach sequence.

Based on work by: Daniel Solomon <daniels@nvidia.com>

BUG=chrome-os-partner:30461
TEST=No screen corruption when switching between dock and undock mode.

Reviewed-on: https://chromium-review.googlesource.com/211177
Commit-Queue: Dylan Reid <dgreid@chromium.org>
Tested-by: Dylan Reid <dgreid@chromium.org>

+++ drivers/video/tegra/dc/dc_priv_defs.h
@@ -182,7 +182,7 @@ struct tegra_dc {
 	u32				vblank_syncpt;
 	u32				*win_syncpt;
 
-	u32				valid_windows;
+	unsigned long int		valid_windows;
 
 	unsigned long			underflow_mask;
 	struct work_struct		reset_work;
